Ensure that the environment is always available when it's needed.
Hide the previous init for the store to ensure that the user either uses Never for Environment and Never for EffectAction or Void for Environment (and then the environment can be set in the initializer).

If the store doesn't ever set the environment, effect actions will wait forever to get executed.

It may be necessary to construct the environment only after the store is constructed because during store construction, the context for the environment is still not there (for example, if the environment depends on the UI or on some other store).

Previously, if the environment wasn't there by the time an effect started, the effect was never executed. With this change, the effect waits until the environment becomes available and executes then. The compiler ensures that the environment is there when it's needed, but the effect may wait forever if the store never sets the environment.

It's impossible to detect when to stop waiting for the environment because the context might become available only in a specific path that the user takes through the app.