Add AsyncResult and store.result()
This is useful for tracking updates to a single property when the updated value might be the same as the previous value. For example, if a user is supposed to provide a value, and there should be an action based on that value (even if the new value is the same as before), store.result() provides a way to get that value.